Quick draft of the programme, ugly but work.
This is just a quick draft of the code, it require to add the alarms directly into the code. i will add the option to add them with user input into a text file that will be read by the program when executed. the goal is to keep it as simple and optimized as possible. i am lazy so i might not finish this project. it would be nice if the user could input a message for each alarm to say with gtts.